refactor: Store completion directory path at install in a variable
The directory in which completion scripts are stored is set dynamically and
stored in COMPLETION_DIR when install.sh is executed, however it could change
between install and uninstall so when scripts/uninstall.sh is executed
COMPLETION_DIR may have a different value. In order to avoid this problem,
install.sh now creates a new variable named AT_INSTALL_COMPLETION_DIR
specifically saying what path was used, and writes this variable into
scripts/config.sh.

When scripts/uninstall.sh is executed, it uses the variable
AT_INSTALL_COMPLETION_DIR from 'scripts/config.sh' in order to unlink the
completion script, and as a cleaning measure, it also removes the line defining
that variable using sed. As a compatibility measure between Linux and Mac OS X
versions of sed, the option -i.bak is used, but as this creates the file
"scripts/config.sh.bak" it is necessary to delete this file.