refactor(install,uninstall): source completion script from .bash_profile
There is not a bash completion directory that works in all systems (e.g. Mac OS,
CentOS and Ubuntu) and that is writable by non-root users, for this reason, a
solution that works in all systems is to write lines into .bash_profile to
source the completion script from there.

The lines written to .bash_profile are deleted by the uninstall.sh script using
sed with the flag -i.bak for compatibility with Mac OS. The resulting .bak is
then removed.

As the only variables in config.sh are COMPLETION_DIR and
AT_INSTALL_COMPLETION_DIR, this change makes config.sh unnecesary.